LibCore: Report error condition when reading process statistics failed

+++ Userland/killall.cpp
@@ -41,8 +41,10 @@ static void print_usage_and_exit()
 static int kill_all(const String& process_name, const unsigned signum)
 {
     auto processes = Core::ProcessStatisticsReader().get_all();
+    if (!processes.has_value())
+        return 1;
 
-    for (auto& it : processes) {
+    for (auto& it : processes.value()) {
         if (it.value.name == process_name) {
             int ret = kill(it.value.pid, signum);
             if (ret < 0)
